+++ bot/meta/Utils/PrintLogs/socketio.py
@@ -1,27 +1,30 @@
-from socketio import Client
-from time import sleep
-
-io = Client()
+# from time import sleep
+from app import io, app
+from status.server_side import serverSide
 
 
 class SocketBot:
     def __init__(self):
         self.connected = False
+        self.first_log = 0
 
     def with_context(self, event: str, data: dict, url: str):
-        url = f"https://{url}"
+        # url = f"https://{url}"
 
         # Verifica se já está conectado antes de tentar se conectar
-        if not self.connected:
-            io.connect(url, namespaces=["/log"], retry=True)
-            self.connected = True
+        # if not self.connected:
+        #     sio.connect(url, namespaces=["/log"], retry=True)
+        #     self.connected = True
 
-        io.emit(event, data, "/log")
-        sleep(1)
+        # sio.emit(event, data, "/log")
+        # sleep(1)
 
         # Após a emissão, desconecta e define o status
-        io.disconnect()
-        self.connected = False
+        # io.disconnect()
+        # self.connected = False
+        data = serverSide(data, data["pid"])
+        with app.app_context():
+            io.emit(event, data, namespace="/log")
 
     def send_message(self, data: dict[str, str | int], url: str):
         try:
